add logout & fix layout
