+++ client/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from './pages/login/login.component';
 import { OauthRedirectComponent } from './pages/oauth-redirect/oauth-redirect.component';
-import { PortalComponent } from './pages/portal/portal.component';
 import { authGuard, guestGuard } from './guards/auth.guard';
+import { PortalComponent } from './pages/portal/portal.component';
 
 const routes: Routes = [
   {
