+++ client/src/app/app.module.ts
@@ -10,7 +10,7 @@ import { NgxSpinnerModule } from "ngx-spinner";
 
 import { ButtonModule } from 'primeng/button';
 import { CardModule } from 'primeng/card';
-import { PortalComponent } from './components/portal/portal.component';
+import { PortalComponent } from './pages/portal/portal.component';
 import { LoginComponent } from './pages/login/login.component';
 import { OauthRedirectComponent } from './pages/oauth-redirect/oauth-redirect.component';
 import { tokenInterceptor } from './config/token.interceptor';
