--- client/src/app/components/portal/portal.component.html
@@ -1,19 +1,0 @@
-<div class="w-full flex gap-3">
-  @for (system of systems; track $index) {
-  <p-card
-    (click)="xRedirect(system)"
-    header="{{ system.name }}"
-    class="cursor-pointer w-3"
-  >
-    <ng-template pTemplate="header">
-      <img alt="{{ system.name }}" src="{{ system.logoUrl }}" />
-    </ng-template>
-
-    <p>{{ system.description }}</p>
-  </p-card>
-  }
-</div>
-
-<button (click)="logout()">Logout</button>
-<button (click)="signInWithGoogle()">Sign in with Google</button>
-<button (click)="testFromSams()">Test from Sams</button>
