+++ client/src/app/pages/portal/portal.component.css
@@ -1,0 +1,13 @@
+.cards-container{
+    display: grid;
+    grid-template-columns: repeat(1, 1fr);
+    gap: 1rem;
+    place-content: center;
+}
+
+@media screen and (min-width: 768px) {
+    .cards-container{
+        grid-template-columns: repeat(3, 1fr);
+        
+    }
+}
